requirements.txt
fastapi
uvicorn
langroid
pyaudio
faster-whisper
numpy
motor
python-dotenv
